remove securepath, simply reject relative paths: .., check / prefix
we could escape the path if for example the base dir was "base" and a
secret directory starts with a name: "base.secret". The request
".secret/secretfile" would translate to "base.secret/secretfile".

reject relative paths: securepath() currently converts paths from
"/../file" to "/.//file" which won't work, so remove it.

always check if a path starts with "/" or is empty.

Signed-off-by: Christoph Lohmann <20h@r-36.net>